A function printing all the elements of a list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -7,7 +7,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t element = 0;
+	int element = 0;
 
 	while (h != NULL)
 	{
